// src/models/IBeer.ts
interface hops {
  name: string;
  amount: {
    value: number;
    unit: string;
  };
  add: string;
  attribute: string;
}

interface malt {
  name: string;
  amount: {
    value: number;
    unit: string;
  };
}

export interface IBeer {
  id: string;
  name: string;
  tagline: string;
  first_brewed: string;
  description: string;
  image_url: string;
  abv: number;
  ibu: number;
  target_fg: number;
  target_og: number;
  ebc: number;
  srm: number;
  ph: number;
  attenuation_level: number;
  volume: {
    value: number;
    unit: string;
  };
  boil_volume: {
    value: number;
    unit: string;
  };
  method: {
    mash_temp: [
      {
        temp: {
          value: number;
          unit: number;
        };
        duration: number;
      }
    ];
    fermentation: {
      temp: {
        value: number;
        unit: string;
      };
    };
    twist: null;
  };
  ingredients: {
    malt: malt[];
    hops: hops[];
    yeast: string;
  };
  food_pairing: [];
  brewers_tips: string;
  contributed_by: string;
}

export interface IDataBeer {
    data: IBeer
}